+++ src/Ext/Techno/Body.h
@@ -125,4 +125,5 @@ public:
 	static void DrawSelfHealPips(TechnoClass* pThis, Point2D* pLocation, RectangleStruct* pBounds);
 	static void ApplyGainedSelfHeal(TechnoClass* pThis);
 	static void SyncIronCurtainStatus(TechnoClass* pFrom, TechnoClass* pTo);
+	static void ApplyMobileRefinery(TechnoClass* pThis);
 };