+++ docs/Whats-New.md
@@ -325,6 +325,7 @@ New:
 - Warhead detonation on all objects on map (by Starkku)
 - Implemented support for PCX images for campaign load screen (by FlyStar)
 - Animated (non-tiberium spawning) TerrainTypes (by Starkku)
+- Mobile Refinery Logic (by NetsuNegi)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)